requirements.txt
python-multipart==0.0.9
dvc==3.50.1
pathspec==0.11.1
scikit-learn==1.3.2
Pillow==10.1.0
tqdm==4.66.1
torch==2.1.2
torchvision==0.16.2
mlflow==2.9.2
fastapi==0.109.0
uvicorn==0.27.0
